requirements.txt
telethon
mautrix
SQLAlchemy
mysqlclient
python-dotenv
rarfile
python-magic